requirements.txt
django
channels==3.0.5
drf-yasg
djangorestframework 
Pillow
djangorestframework-simplejwt
django-cors-headers